backend/requirements.txt
flask
python-dotenv
beautifulsoup4
requests
selenium
google-generativeai
webdriver-manager